+++ packages/react-native-messaging-core/ios/Sources/NablaMessagingClient/NablaMessagingClientModule.swift
@@ -10,6 +10,7 @@ final class NablaMessagingClientModule: NSObject {
     private var sendMessageCancellable: Cancellable?
     private var deleteMessageCancellable: Cancellable?
     private var markConversationAsSeenCancellable: Cancellable?
+    private var setIsTypingCancellable: Cancellable?
 
     @objc(createConversation:providerIds:callback:)
     func createConversation(
@@ -117,6 +118,32 @@ final class NablaMessagingClientModule: NSObject {
                 callback([error.dictionaryRepresentation])
             }
         })
+    }
+
+    @objc(setIsTyping:conversationId:callback:)
+    func setIsTyping(
+        _ isTyping: Bool,
+        conversationIdMap: [String: Any],
+        callback: @escaping RCTResponseSenderBlock
+    ) {
+        guard let conversationId = conversationIdMap.asConversationId else {
+            callback([
+                InternalError.createDictionaryRepresentation(message: "Bad conversationId `\(conversationIdMap)`")
+            ])
+            return
+        }
+
+        setIsTypingCancellable = NablaMessagingClient.shared.setIsTyping(
+            isTyping,
+            inConversationWithId: conversationId,
+            handler: { result in
+                switch result {
+                case .success:
+                    callback([NSNull()])
+                case .failure(let error):
+                    callback([error.dictionaryRepresentation])
+                }
+            })
     }
 
     @objc class func requiresMainQueueSetup() -> Bool {