+++ packages/react-native-messaging-core/src/NablaMessagingClient/NativeModules/index.ts
@@ -2,3 +2,4 @@ export * from './ConversationItemsWatcherModule';
 export * from './ConversationListWatcherModule';
 export * from './ConversationWatcherModule';
 export * from './NablaMessagingClientModule';
+export * from './Callback';